requirements_ocr.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
